replaced MazeState with MazeModel class
because it makes more sense
